chore: Change mutation chance from float to double in DivideCommand
Updated DivideCommand to use double instead of float for mutation chance for improved precision. Adjusted related code in constructor, mutation distribution, and command registration to reflect this change. Also improved command name formatting to show mutation chance with two decimal places.

+++ include/CellularSimulator/Core/Commands/DivideCommand.h
@@ -13,7 +13,7 @@ namespace Core
 class DivideCommand : public Command
 {
 public:
-    DivideCommand(float InMutationChance);
+    DivideCommand(double InMutationChance);
     
     /**
      * @brief Executes the divide command to create offspring cell
@@ -33,7 +33,7 @@ public:
     void Execute(Simulator& Sim, Cell& Agent) override;
 
 private:
-    float MutationChance;
+    double MutationChance;
 };
 
 }  // namespace Core
